EngineSync: Remove BaseSyncableListener class
