+++ src/engine/sync/enginesync.cpp
@@ -1,5 +1,6 @@
 #include "engine/sync/enginesync.h"
 
+#include <QMetaType>
 #include <QStringList>
 
 #include "engine/channels/enginechannel.h"
@@ -10,13 +11,21 @@
 
 namespace {
 const mixxx::Logger kLogger("EngineSync");
-} // namespace
+const QString kInternalClockGroup = QStringLiteral("[InternalClock]");
+} // anonymous namespace
 
 EngineSync::EngineSync(UserSettingsPointer pConfig)
-        : BaseSyncableListener(pConfig) {
+        : m_pConfig(pConfig),
+          m_pInternalClock(new InternalClock(kInternalClockGroup, this)),
+          m_pMasterSyncable(nullptr) {
+    qRegisterMetaType<SyncMode>("SyncMode");
+    m_pInternalClock->setMasterBpm(124.0);
 }
 
 EngineSync::~EngineSync() {
+    // We use the slider value because that is never set to 0.0.
+    m_pConfig->set(ConfigKey(kInternalClockGroup, "bpm"), ConfigValue(m_pInternalClock->getBpm()));
+    delete m_pInternalClock;
 }
 
 Syncable* EngineSync::pickMaster(Syncable* enabling_syncable) {
@@ -477,4 +486,139 @@ bool EngineSync::otherSyncedPlaying(const QString& group) {
         }
     }
     return othersInSync;
+}
+
+void EngineSync::addSyncableDeck(Syncable* pSyncable) {
+    if (m_syncables.contains(pSyncable)) {
+        qDebug() << "EngineSync: already has" << pSyncable;
+        return;
+    }
+    m_syncables.append(pSyncable);
+}
+
+void EngineSync::onCallbackStart(int sampleRate, int bufferSize) {
+    m_pInternalClock->onCallbackStart(sampleRate, bufferSize);
+}
+
+void EngineSync::onCallbackEnd(int sampleRate, int bufferSize) {
+    m_pInternalClock->onCallbackEnd(sampleRate, bufferSize);
+}
+
+EngineChannel* EngineSync::getMaster() const {
+    return m_pMasterSyncable ? m_pMasterSyncable->getChannel() : nullptr;
+}
+
+Syncable* EngineSync::getSyncableForGroup(const QString& group) {
+    foreach (Syncable* pSyncable, m_syncables) {
+        if (pSyncable->getGroup() == group) {
+            return pSyncable;
+        }
+    }
+    return nullptr;
+}
+
+bool EngineSync::syncDeckExists() const {
+    for (const auto& pSyncable : qAsConst(m_syncables)) {
+        if (pSyncable->isSynchronized() && pSyncable->getBaseBpm() > 0) {
+            return true;
+        }
+    }
+    return false;
+}
+
+double EngineSync::masterBpm() const {
+    if (m_pMasterSyncable) {
+        return m_pMasterSyncable->getBpm();
+    }
+    return m_pInternalClock->getBpm();
+}
+
+double EngineSync::masterBeatDistance() const {
+    if (m_pMasterSyncable) {
+        return m_pMasterSyncable->getBeatDistance();
+    }
+    return m_pInternalClock->getBeatDistance();
+}
+
+double EngineSync::masterBaseBpm() const {
+    if (m_pMasterSyncable) {
+        return m_pMasterSyncable->getBaseBpm();
+    }
+    return m_pInternalClock->getBaseBpm();
+}
+
+void EngineSync::setMasterBpm(Syncable* pSource, double bpm) {
+    //qDebug() << "EngineSync::setMasterBpm" << pSource << bpm;
+    if (pSource != m_pInternalClock) {
+        m_pInternalClock->setMasterBpm(bpm);
+    }
+    foreach (Syncable* pSyncable, m_syncables) {
+        if (pSyncable == pSource ||
+                !pSyncable->isSynchronized()) {
+            continue;
+        }
+        pSyncable->setMasterBpm(bpm);
+    }
+}
+
+void EngineSync::setMasterInstantaneousBpm(Syncable* pSource, double bpm) {
+    if (pSource != m_pInternalClock) {
+        m_pInternalClock->setInstantaneousBpm(bpm);
+    }
+    foreach (Syncable* pSyncable, m_syncables) {
+        if (pSyncable == pSource ||
+                !pSyncable->isSynchronized()) {
+            continue;
+        }
+        pSyncable->setInstantaneousBpm(bpm);
+    }
+}
+
+void EngineSync::setMasterBeatDistance(Syncable* pSource, double beatDistance) {
+    if (pSource != m_pInternalClock) {
+        m_pInternalClock->setMasterBeatDistance(beatDistance);
+    }
+    foreach (Syncable* pSyncable, m_syncables) {
+        if (pSyncable == pSource ||
+                !pSyncable->isSynchronized()) {
+            continue;
+        }
+        pSyncable->setMasterBeatDistance(beatDistance);
+    }
+}
+
+void EngineSync::setMasterParams(
+        Syncable* pSource, double beatDistance, double baseBpm, double bpm) {
+    //qDebug() << "EngineSync::setMasterParams, source is" << pSource->getGroup() << beatDistance << baseBpm << bpm;
+    if (pSource != m_pInternalClock) {
+        m_pInternalClock->setMasterParams(beatDistance, baseBpm, bpm);
+    }
+    foreach (Syncable* pSyncable, m_syncables) {
+        if (pSyncable == pSource ||
+                !pSyncable->isSynchronized()) {
+            continue;
+        }
+        pSyncable->setMasterParams(beatDistance, baseBpm, bpm);
+    }
+}
+
+void EngineSync::checkUniquePlayingSyncable() {
+    int playing_sync_decks = 0;
+    Syncable* unique_syncable = nullptr;
+    foreach (Syncable* pSyncable, m_syncables) {
+        if (!pSyncable->isSynchronized()) {
+            continue;
+        }
+
+        if (pSyncable->isPlaying()) {
+            if (playing_sync_decks > 0) {
+                return;
+            }
+            unique_syncable = pSyncable;
+            ++playing_sync_decks;
+        }
+    }
+    if (playing_sync_decks == 1) {
+        unique_syncable->notifyOnlyPlayingSyncable();
+    }
 }
