+++ src/engine/sync/enginesync.h
@@ -2,13 +2,15 @@
 
 #include "preferences/usersettings.h"
 #include "engine/sync/syncable.h"
-#include "engine/sync/basesyncablelistener.h"
+
+class InternalClock;
+class EngineChannel;
 
 /// EngineSync is the heart of the Mixxx Master Sync engine.  It knows which objects
 /// (Decks, Internal Clock, etc) are participating in Sync and what their statuses
 /// are. It also orchestrates sync handoffs between different decks as they play,
 /// stop, or request their status to change.
-class EngineSync : public BaseSyncableListener {
+class EngineSync : public SyncableListener {
   public:
     explicit EngineSync(UserSettingsPointer pConfig);
     ~EngineSync() override;
@@ -53,6 +55,17 @@ class EngineSync : public BaseSyncableListener {
     // of other Syncables that are playing
     bool otherSyncedPlaying(const QString& group);
 
+    void addSyncableDeck(Syncable* pSyncable);
+    EngineChannel* getMaster() const;
+    void onCallbackStart(int sampleRate, int bufferSize);
+    void onCallbackEnd(int sampleRate, int bufferSize);
+
+    // Only for testing. Do not use.
+    Syncable* getSyncableForGroup(const QString& group);
+    Syncable* getMasterSyncable() override {
+        return m_pMasterSyncable;
+    }
+
   private:
     // Iterate over decks, and based on sync and play status, pick a new master.
     // if enabling_syncable is not null, we treat it as if it were enabled because we may
@@ -75,4 +88,44 @@ class EngineSync : public BaseSyncableListener {
 
     // Unsets all sync state on a Syncable.
     void deactivateSync(Syncable* pSyncable);
+
+    // This utility method returns true if it finds a deck not in SYNC_NONE mode.
+    bool syncDeckExists() const;
+
+    // Return the current BPM of the master Syncable. If no master syncable is
+    // set then returns the BPM of the internal clock.
+    double masterBpm() const;
+
+    // Returns the current beat distance of the master Syncable. If no master
+    // Syncable is set, then returns the beat distance of the internal clock.
+    double masterBeatDistance() const;
+
+    // Returns the overall average BPM of the master Syncable if it were playing
+    // at 1.0 rate. This is used to calculate half/double multipliers and whether
+    // the master has a bpm at all.
+    double masterBaseBpm() const;
+
+    // Set the BPM on every sync-enabled Syncable except pSource.
+    void setMasterBpm(Syncable* pSource, double bpm);
+
+    // Set the master instantaneous BPM on every sync-enabled Syncable except
+    // pSource.
+    void setMasterInstantaneousBpm(Syncable* pSource, double bpm);
+
+    // Set the master beat distance on every sync-enabled Syncable except
+    // pSource.
+    void setMasterBeatDistance(Syncable* pSource, double beatDistance);
+
+    void setMasterParams(Syncable* pSource, double beatDistance, double baseBpm, double bpm);
+
+    // Check if there is only one playing syncable deck, and notify it if so.
+    void checkUniquePlayingSyncable();
+
+    UserSettingsPointer m_pConfig;
+    // The InternalClock syncable.
+    InternalClock* m_pInternalClock;
+    // The current Syncable that is the master.
+    Syncable* m_pMasterSyncable;
+    // The list of all Syncables registered via addSyncableDeck.
+    QList<Syncable*> m_syncables;
 };
